Added requirements file (#72)
This is great - thanks! This is much cleaner. 

I use a variety of python versions to develop - 3.10.12 on my desktop with WSL, whatever is on the pi and on my steamdeck. 

Thanks for adding the windows part - the real challenge on windows will be `python-mpv-jsonipc` which is theoretically possible, but I've never went through the steps.