Series sequences (#159)
* Basic working squences.

* check for catalog version on load

* Framework for handling schedule deletions

* use liquid manager for resetting schedules

* preserves sequence when schedules are deleted

When the schedule adds a show from a sequenced series, it will store a reference to the sequence with the schedule entry. When a schedule is to be deleted, it will find the first instance of each sequence after the current timestamp (meaning, the first one that was scheduled but hasn't been played yet) and tell the sequence to reset to that point. (anything in the past is considered 'played' - even if no-one actually watched it)