Web Type Stations (#371)
## 🌐 Web Channel
This is a channel type that uses a web browser component to display a web page - essentially turning any URL into a TV station. To show how this works, we'll make a diagnostics channel based on the FieldStation42 built-in `static/diagnostics.html` that displays system information like memory and cpu usage.

### 📊 Diagnostic Channel example
The following configuration will create a diagnostics channel:
```json
{"station_conf" : {
    "network_name" : "diagnostic",
    "network_type" : "web",
    "channel_number": 7,
    "web_url": "http://localhost:4242/static/diagnostics.html"
}}
```
To see what the channel will look like, you can visit the `web_url` in your web browser while the player is running.

## 🌤️ WeatherStar Channel
To implement the weather channel, we will use `"network_type": "web"` configuration where we point the `web_url` to point to a WeatherStar installation. For this example, we will use [WeatherStar 4000+](https://github.com/netbymatt/ws4kp), but the [international version](https://github.com/mwood77/ws4kp-international) will work as well. 

### Install WeatherStar Docker
Using portainer, go to stacks and enter the following:

```yaml
version: '3.8'
services:
  weatherstar4000:
    image: ghcr.io/netbymatt/ws4kp
    container_name: ws4kp
    ports:
      - "9090:8080"
    environment:
      - WSQS_latLonQuery=Seattle WA USA
    restart: unless-stopped
```
Note on ports: the default is `8080:8080`, but since 8080 is often already taken on busy server, I changed it to 9090 by setting port mapping to `9090:8080`

### Configuring the URL
After deploying, visit the site in your web browser. Select all the options you want in your display in the check boxes and configuration. Then select `Copy Permalink` which will give you something like this:
```
http://<IP_ADDRESS_HERE>:9090/index.html?hazards-checkbox=true&current-weather-checkbox=true&latest-observations-checkbox=true&hourly-checkbox=true&hourly-graph-checkbox=true&travel-checkbox=true&regional-forecast-checkbox=true&local-forecast-checkbox=true&extended-forecast-checkbox=true&almanac-checkbox=true&spc-outlook-checkbox=true&radar-checkbox=true&settings-wide-checkbox=false&settings-kiosk-checkbox=false&settings-stickyKiosk-checkbox=false&settings-scanLines-checkbox=false&settings-customFeedEnable-checkbox=false&settings-speed-select=1.00&settings-scanLineMode-select=auto&settings-units-select=us&settings-mediaVolume-select=0.75&txtLocation=Seattle+WA+USA&settings-customFeed-string=&share-link-url=&latLonQuery=Seattle%2C+WA%2C+USA&latLon=%7B%22lat%22%3A47.6032%2C%22lon%22%3A-122.3302%7D
```
To make the URL open in kiosk mode and to make it start autoplaying music, you will need to add the following to the end of the URL:
```
&kiosk=true&settings-mediaPlaying-boolean=true
```

### Configuring the Weather Channel
Configure the channel to use a `network_type` of `web` and set the `web_url` parameter to be the URL to your WeatherStar4K installation.
```
{"station_conf" : {
    "network_name" : "WeatherTV",
    "network_type" : "web",
    "channel_number": 14,
    "web_url": "http://<IP_ADDRESS>:9090/<LONG_URL_PARAMS>&kiosk=true&settings-mediaPlaying-boolean=true"
}}
```